3d-function-8.tex: convert argument of sin() to degrees.
tikz apparently expectes the argument to sin() in degrees. This is in
strong contrast to the usual conventions in math, where the argument is
expected in radians.

With this example it actually provides a bogus graph - it actually
should approach f(x,y) = 1 for large x and y (since 1/(x^2+y^2) is
small for large x,y; hence sin(1/(x^2+y^2)) can be approximated by
1/(x^2+y^2), resulting in f(x,y) -> 1 for large x,y).

Also I've adjusted the plot range a bit, since this then actually shows
the "interesting" center of the function (x=y=0).